+++ application/comit_node/src/swap_protocols/rfc003/actions/bob/btc_erc20.rs
@@ -83,6 +83,7 @@ impl StateActions for SwapStates<Bob<Bitcoin, Ethereum, BitcoinQuantity, Erc20Qu
     type Redeem = bitcoin::SpendOutput;
     type Refund = ethereum::SendTransaction;
 
+    #[allow(clippy::type_complexity)]
     fn actions(
         &self,
     ) -> Vec<
