+++ application/comit_node/src/swap_protocols/rfc003/actions/mod.rs
@@ -35,6 +35,7 @@ pub trait StateActions {
     type Redeem;
     type Refund;
 
+    #[allow(clippy::type_complexity)]
     fn actions(
         &self,
     ) -> Vec<
