+++ application/comit_node/src/swap_protocols/rfc003/ethereum/erc20_htlc.rs
@@ -59,7 +59,7 @@ impl Erc20Htlc {
     pub fn funding_tx_payload(&self, htlc_contract_address: Address) -> Bytes {
         let transfer_fn_abi = base16!("A9059CBB");
         let htlc_contract_address = <[u8; 20]>::from(htlc_contract_address);
-        let amount = <[u8; 32]>::from(self.amount.clone());
+        let amount = <[u8; 32]>::from(self.amount);
 
         let mut data = [0u8; 4 + 32 + 32];
         data[..4].copy_from_slice(transfer_fn_abi);
